Complete CoinBanner separate component
