+++ src/Pages/Landing/index.js
@@ -1,7 +1,7 @@
 import React, { useEffect } from "react";
 import { useDispatch } from "react-redux";
 import Banner from "./Banner";
-import CoinBanner from "./CoinBanner";
+import CoinBanner from "./CoinBanner/index";
 import News from "./News";
 import { updatePageName } from "../../Redux/Actions/actionCreator";
